+++ frontend/src/views/failures/details/index.js
